requirements.txt
Flask==3.0.3
gunicorn==21.2.0
Werkzeug==3.0.3
Jinja2==3.1.4
numpy==1.26.4
Pillow==10.3.0
tqdm==4.66.4
opencv-python-headless==4.9.0.80
scikit-learn==1.4.0
joblib==1.4.2
threadpoolctl==3.5.0
scipy==1.13.1
matplotlib==3.8.4
requests==2.32.2
onnxruntime==1.17.0
tensorflow==2.16.0